+++ apps/frontend/src/app/3_organisms/RuneBridgeDialog/components/SendFlow/components/AmountForm.tsx
@@ -27,14 +27,14 @@ import {
   WITHDRAW_MIN_AMOUNT,
   WITHDRAW_MAX_AMOUNT,
 } from '../../../constants';
-import { useContractContext } from '../../../contexts/contract';
+import { useRuneContext } from '../../../contexts/rune';
 import { SendFlowContext, SendFlowStep } from '../../../contexts/sendflow';
 import { useRuneBridgeLocked } from '../../../hooks/useRuneBridgeLocked';
 import { TransferPolicies } from '../../TransferPolicies';
 
 export const AmountForm: React.FC = () => {
   const { amount, limits, selectedToken, set } = useContext(SendFlowContext);
-  const { tokenBalances } = useContractContext();
+  const { tokenBalances } = useRuneContext();
   const runeBridgeLocked = useRuneBridgeLocked();
 
   const [value, setValue] = useState(amount || '');
